Fixed avatar action visualization
