+++ public/index.html
@@ -21,7 +21,7 @@
 
 
 
-<script type="module" src="script.js"></script>
+<script type="module" src="script3.js"></script>
 </body>
 
 </html>
