+++ src/Scuti.ts
@@ -3,8 +3,7 @@ import { IRendererConfiguration } from "./interfaces/Configuration.interface";
 import { Logger } from "./utilities/Logger";
 import { PixiPlugin } from "gsap/PixiPlugin";
 import { gsap } from "gsap";
-import {Stage} from "@pixi/layers";
-import {handleClick} from "./objects/interactions/HitInteraction";
+import { Stage } from "@pixi/layers";
 
 /**
  * Convenience class to create a new Scuti renderer.
